Split the module program up into a base section (available in libflprapp) and a extensible instantiation section.
